+++ front/app/components/SeatBasedBilling/AddModeratorsModal/index.tsx
@@ -3,8 +3,8 @@ import React, { useState } from 'react';
 // Components
 import { Box, Button, Text } from '@citizenlab/cl2-component-library';
 import Modal from 'components/UI/Modal';
-import SeatInfo from 'components/SeatInfo';
-import SeatSetSuccess from 'components/admin/SeatSetSuccess';
+import SeatInfo from 'components/SeatBasedBilling/SeatInfo';
+import SeatSetSuccess from 'components/SeatBasedBilling/SeatSetSuccess';
 
 // Translation
 import { FormattedMessage, useIntl } from 'utils/cl-intl';
@@ -16,8 +16,8 @@ import useSeats from 'api/seats/useSeats';
 import useFeatureFlag from 'hooks/useFeatureFlag';
 
 // Utils
-import { getExceededLimitInfo } from 'components/SeatInfo/utils';
-import BillingWarning from 'components/SeatInfo/BillingWarning';
+import { getExceededLimitInfo } from 'components/SeatBasedBilling/utils';
+import BillingWarning from 'components/SeatBasedBilling/BillingWarning';
 
 interface Props {
   showModal: boolean;
