+++ front/app/components/SeatBasedBilling/SeatInfo/index.tsx
