+++ front/app/components/SeatBasedBilling/SeatInfo/messages.ts
