+++ front/app/containers/Admin/invitations/invite/index.tsx
@@ -24,9 +24,9 @@ import HelmetIntl from 'components/HelmetIntl';
 import Button from 'components/UI/Button';
 import Warning from 'components/UI/Warning';
 const InviteUsersWithSeatsModal = lazy(
-  () => import('components/admin/InviteUsersWithSeatsModal')
+  () => import('components/SeatBasedBilling/InviteUsersWithSeatsModal')
 );
-import SeatInfo from 'components/SeatInfo';
+import SeatInfo from 'components/SeatBasedBilling/SeatInfo';
 import useFeatureFlag from 'hooks/useFeatureFlag';
 
 // services
